Make `getValue` an external function that can be passed in
The way that values are aggregated from peers is highly dependent on use case. You might wish to add the values together, or multiply them, or something else entirely. Point being, it doesn't make sense to try to couple this with the metric. The metric determines how *confidence* is aggregated, not value.