Differentiate between atheists and agnostics
An atheist says "there is no god", while an agnostic simply does not say "there is a god". These are two very different things. It applies here because I was previously treating the statement, "peer A has cred value of 0" the same as "I have nothing to say of peer A", which meant that going from not saying anything to saying just a little bit could produce unexpected results.