Experiment with techniques to optimize performance
Techniques being used here are (1) caching and (2) being apathetic to graph cycles (which actually just enables more caching).

Previously I was tracking paths when searching through the graph, having each recursive call know previously "visited" nodes and thus be able to stop when visiting a repeat node. This is problematic for caching as every single call is now unique, and values cannot be cached (the `visited` array will always be unique).

But without a method for preventing cycles, then the previous algorithm would encounter an infinite loop upon reaching a cycle. Thus I adjusted the algorithm to instead use a pagerank-style approach to deal with cycles. That is, recursion is abandoned in favor of the concept of *iterations*. In each iteration, a node requests information from its peers (as in the old algorithm). But each peer responds not by recursively calling its own peers, but rather by simply returning its locally cached values. Each peer only updates its cached values on its turn in the iteration, basing its updated value on the cached values of its peers. In the first iteration this may not be a very useful result. But as the number of iterations increases, the cached values will stabilize; and at a relatively low number they should be "good enough" approximations of the true equilibrium values (when the number of iterations is infinity).